services/tracking/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
paho-mqtt==1.6.1
SQLAlchemy==2.0.32


